requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==3.2
django-heroku==0.3.1
gunicorn==20.1.0
numpy==1.22.2
Pillow==9.0.1
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
torch==1.10.2
torchvision==0.11.3
typing_extensions==4.1.1
whitenoise==6.0.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.7.1+cpu